// src/app/common/data-core/enums/road/road-object/road-object-type.enum.ts
/**	RoadObjectType (部件类型)	*/
export enum RoadObjectType {
  /**	消防栓	*/
  FireHydrant = 1,
  /**	公交车站	*/
  BusStation = 2,
  /**	垃圾桶	*/
  TrashCan = 3,
  /**	重要通道	*/
  Passage = 4,
  /**	电话亭	*/
  TelephoneBooth = 5,
  /** 机非隔离栏 */
  CycleLaneSeparator = 6,
  NoParking = 7,
  PublicityWall = 8,
}